Intento de separar lógica de login
