+++ src/App.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable no-unused-vars */
 import React from 'react'
-import { Route, Routes, createBrowserRouter } from 'react-router-dom';
+import { createBrowserRouter } from 'react-router-dom';
 import './App.css'
 
 import Home from './components/pages/Home/home.jsx';
